% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{create_island_spec}
\alias{create_island_spec}
\title{create island spec helper function}
\usage{
create_island_spec(
  time,
  mainland_n,
  K,
  immig_rate,
  ext_rate,
  ana_rate,
  clado_rate,
  immig_rate2,
  ext_rate2,
  ana_rate2,
  clado_rate2,
  trans_rate,
  trans_rate2,
  M2
)
}
\arguments{
\item{time}{time}

\item{mainland_n}{mainland n}

\item{K}{K}

\item{immig_rate}{igr}

\item{ext_rate}{er}

\item{ana_rate}{ar}

\item{clado_rate}{cr}

\item{immig_rate2}{igr2}

\item{ext_rate2}{er2}

\item{ana_rate2}{ar2}

\item{clado_rate2}{cr2}

\item{trans_rate}{tr}

\item{trans_rate2}{tr2}

\item{M2}{m2}
}
\value{
stochastic island spec table
}
\description{
create island spec helper function
}
