% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{DAISIE_test_execute_extinction}
\alias{DAISIE_test_execute_extinction}
\title{testing function}
\usage{
DAISIE_test_execute_extinction(island_spec, extinct)
}
\arguments{
\item{island_spec}{island_spec}

\item{extinct}{index}
}
\value{
island spec
}
\description{
testing function
}
