% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{execute_time_loop}
\alias{execute_time_loop}
\title{CPP function to execute expensive loop}
\usage{
execute_time_loop(
  timeval,
  total_time,
  gam,
  laa,
  lac,
  mu,
  area_pars_from_R,
  K,
  mainland_n,
  trait_pars_R
)
}
\arguments{
\item{timeval}{timeval}

\item{total_time}{total time}

\item{gam}{immigration rate}

\item{laa}{anagenetic rate}

\item{lac}{cladogenetic rate}

\item{mu}{extinction rate}

\item{area_pars_from_R}{area pars}

\item{K}{K}

\item{mainland_n}{number of species on mainland}

\item{trait_pars_R}{trait pars}
}
\value{
list with stt_table and island_spec
}
\description{
CPP function to execute expensive loop
}
