% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_simulation_outputs}
\alias{is_simulation_outputs}
\title{Measures if the input is a valid collection of simulation
outputs.}
\usage{
is_simulation_outputs(simulation_outputs)
}
\arguments{
\item{simulation_outputs}{A list with matrices and vectors of simulation
produced by DAISIE_sim functions.}
}
\value{
TRUE if the input is a valid collection of simulation
outputs.
}
\description{
Measures if the input is a valid collection of simulation
outputs.
}
\author{
Richel J.C Bilderbeek, Pedro Neves
}
\keyword{internal}
