% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_event}
\alias{sample_event}
\title{function to draw event.}
\usage{
sample_event(event_prob)
}
\arguments{
\item{event_prob}{event probability vector}
}
\value{
event
}
\description{
function to draw event.
}
