% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_anagenesis}
\alias{test_anagenesis}
\title{test execute anagenesis}
\usage{
test_anagenesis(island_spec_R, maxspecID, focal_trait)
}
\arguments{
\item{island_spec_R}{island spec table}

\item{maxspecID}{maxspecid}

\item{focal_trait}{focaltrait}
}
\value{
island spec table
}
\description{
test execute anagenesis
}
