% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_cladogenesis}
\alias{test_cladogenesis}
\title{test execute cladogenesis}
\usage{
test_cladogenesis(island_spec_R, timeval, maxspecID, focal_trait)
}
\arguments{
\item{island_spec_R}{island spec table}

\item{timeval}{timeval}

\item{maxspecID}{maxspecid}

\item{focal_trait}{focaltrait}
}
\value{
island spec table
}
\description{
test execute cladogenesis
}
