% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_execute_extinction}
\alias{test_execute_extinction}
\title{test execute extinction without rng}
\usage{
test_execute_extinction(island_spec_R, index)
}
\arguments{
\item{island_spec_R}{island spec table}

\item{index}{index}
}
\value{
island spec table
}
\description{
test execute extinction without rng
}
