% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_extinction}
\alias{test_extinction}
\title{test extinction}
\usage{
test_extinction(island_spec_R, focal_trait)
}
\arguments{
\item{island_spec_R}{island spec table}

\item{focal_trait}{focal trait}
}
\value{
island spec table
}
\description{
test extinction
}
