% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_get_ana_rate}
\alias{test_get_ana_rate}
\title{function to test get_ana_rate}
\usage{
test_get_ana_rate(
  laa,
  num_immigrants,
  ana_rate2,
  num_spec_trait1,
  num_spec_trait2
)
}
\arguments{
\item{laa}{anagenesis rate trait 1}

\item{num_immigrants}{number of immigrants}

\item{ana_rate2}{ana_rate trait 2}

\item{num_spec_trait1}{num_spec trait 1}

\item{num_spec_trait2}{num_spec trait 2}
}
\value{
two rates
}
\description{
function to test get_ana_rate
}
