% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_get_clado_rate}
\alias{test_get_clado_rate}
\title{function to test get_ext_rate}
\usage{
test_get_clado_rate(
  lac,
  num_spec,
  K,
  A,
  clado_rate2,
  num_spec_trait1,
  num_spec_trait2
)
}
\arguments{
\item{lac}{caldo rate trait 1}

\item{num_spec}{number of immigrants}

\item{K}{K}

\item{A}{A}

\item{clado_rate2}{clado rate trait 2}

\item{num_spec_trait1}{num_spec trait 1}

\item{num_spec_trait2}{num_spec trait 2}
}
\value{
two rates
}
\description{
function to test get_ext_rate
}
