% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_get_ext_rate}
\alias{test_get_ext_rate}
\title{function to test get_ext_rate}
\usage{
test_get_ext_rate(mu, num_spec, A, ext_rate2, num_spec_trait1, num_spec_trait2)
}
\arguments{
\item{mu}{mu}

\item{num_spec}{num species}

\item{A}{A}

\item{ext_rate2}{ext_rate2}

\item{num_spec_trait1}{num_spec trait 1}

\item{num_spec_trait2}{num_spec trait 2}
}
\value{
two rates
}
\description{
function to test get_ext_rate
}
