% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_get_immig_rate}
\alias{test_get_immig_rate}
\title{function to test get_immigration_rate}
\usage{
test_get_immig_rate(gam, A, num_spec, K, mainland_n, mainland_n2, immig_rate2)
}
\arguments{
\item{gam}{gam}

\item{A}{A}

\item{num_spec}{num species}

\item{K}{K value}

\item{mainland_n}{num mainland species trait 1}

\item{mainland_n2}{num mainland species trait 2}

\item{immig_rate2}{gam2}
}
\value{
two rates
}
\description{
function to test get_immigration_rate
}
