% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_get_trans_rate}
\alias{test_get_trans_rate}
\title{function to test get_trans_rate}
\usage{
test_get_trans_rate(trans_rate, trans_rate2, num_spec_trait1, num_spec_trait2)
}
\arguments{
\item{trans_rate}{tr1}

\item{trans_rate2}{tr2}

\item{num_spec_trait1}{num_spec trait 1}

\item{num_spec_trait2}{num_spec trait 2}
}
\value{
two rates
}
\description{
function to test get_trans_rate
}
