% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_immigration2}
\alias{test_immigration2}
\title{test immigration2}
\usage{
test_immigration2(island_spec_R, mainland_spec_R, timeval, M2)
}
\arguments{
\item{island_spec_R}{island spec table}

\item{mainland_spec_R}{mainland species vector}

\item{timeval}{timeval}

\item{M2}{M2}
}
\value{
island spec table
}
\description{
test immigration2
}
