% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_transition}
\alias{test_transition}
\title{test transition}
\usage{
test_transition(island_spec_R, focal_trait)
}
\arguments{
\item{island_spec_R}{island spec table}

\item{focal_trait}{focaltrait}
}
\value{
island spec table
}
\description{
test transition
}
