% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_update_rates}
\alias{test_update_rates}
\title{function to test get_trans_rate}
\usage{
test_update_rates(
  timeval,
  total_time,
  gam,
  laa,
  lac,
  mu,
  K,
  num_spec,
  num_immigrants,
  mainland_n,
  island_spec_R,
  trait_pars_R
)
}
\arguments{
\item{timeval}{timeval}

\item{total_time}{total time}

\item{gam}{gam}

\item{laa}{laa}

\item{lac}{lac}

\item{mu}{mu}

\item{K}{K}

\item{num_spec}{num_spec}

\item{num_immigrants}{num_immigrants}

\item{mainland_n}{mainland_n}

\item{island_spec}{island_spec}

\item{trait_pars}{trait pars}
}
\value{
two rates
}
\description{
function to test get_trans_rate
}
