% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_immigration}
\alias{test_immigration}
\title{test immigration}
\usage{
test_immigration(timeval, mainland_spec_R, island_spec_R)
}
\arguments{
\item{timeval}{timeval}

\item{island_spec_R}{matrix}

\item{mainland_spec}{vector}
}
\value{
stringmatrix island_spec
}
\description{
test immigration
}
