% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_sample_spec}
\alias{test_sample_spec}
\title{test sample spec}
\usage{
test_sample_spec(mainland_spec_R, M2)
}
\arguments{
\item{mainland_spec_R}{mainland spec table}

\item{M2}{number species of trait 2 on mainland}
}
\value{
species index
}
\description{
test sample spec
}
